Factor out test utils into a new trait called TestUtils.
- Mark the LiabilitySugar class as deprecated.
