+++ test/uk/gov/hmrc/selfassessmentapi/repositories/domain/builders/SelfEmploymentBuilder.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.selfassessmentapi.repositories.domain.builders
 
 import reactivemongo.bson.BSONObjectID
-import uk.gov.hmrc.selfassessmentapi.SelfAssessmentSugar._
+import uk.gov.hmrc.selfassessmentapi.TestUtils._
 import uk.gov.hmrc.selfassessmentapi.controllers.api.selfemployment.{SelfEmployment => _, _}
 import uk.gov.hmrc.selfassessmentapi.repositories.domain._
