+++ test/uk/gov/hmrc/selfassessmentapi/repositories/domain/builders/UnearnedIncomeBuilder.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.selfassessmentapi.repositories.domain.builders
 
 import reactivemongo.bson.BSONObjectID
-import uk.gov.hmrc.selfassessmentapi.SelfAssessmentSugar._
+import uk.gov.hmrc.selfassessmentapi.TestUtils._
 import uk.gov.hmrc.selfassessmentapi.repositories.domain.{UnearnedIncome, UnearnedIncomesDividendSummary, UnearnedIncomesSavingsIncomeSummary}
 
 case class UnearnedIncomeBuilder(objectID: BSONObjectID = BSONObjectID.generate) {
